+++ packages/asset-service/src/generateAssetData/baseAssets.ts
@@ -1,12 +1,13 @@
+import * as caip from '@shapeshiftoss/caip'
 import { Asset } from '@shapeshiftoss/types'
 
 export const ethereum: Asset = {
-  assetId: 'eip155:1/slip44:60',
-  chainId: 'eip155:1',
+  assetId: caip.ethAssetId,
+  chainId: caip.ethChainId,
   symbol: 'ETH',
   name: 'Ethereum',
   precision: 18,
-  color: '#FFFFFF',
+  color: '#5C6BC0',
   icon: 'https://assets.coincap.io/assets/icons/256/eth.png',
   explorer: 'https://etherscan.io',
   explorerAddressLink: 'https://etherscan.io/address/',
@@ -14,12 +15,12 @@ export const ethereum: Asset = {
 }
 
 export const bitcoin: Asset = {
-  assetId: 'bip122:000000000019d6689c085ae165831e93/slip44:0',
-  chainId: 'bip122:000000000019d6689c085ae165831e93',
+  assetId: caip.btcAssetId,
+  chainId: caip.btcChainId,
   symbol: 'BTC',
   name: 'Bitcoin',
   precision: 8,
-  color: '#FFFFFF',
+  color: '#FF9800',
   icon: 'https://assets.coincap.io/assets/icons/256/btc.png',
   explorer: 'https://live.blockcypher.com',
   explorerAddressLink: 'https://live.blockcypher.com/btc/address/',
@@ -32,7 +33,7 @@ export const tBitcoin: Asset = {
   symbol: 'BTC',
   name: 'Bitcoin',
   precision: 8,
-  color: '#FFFFFF',
+  color: '#FF9800',
   icon: 'https://assets.coincap.io/assets/icons/256/btc.png',
   explorer: 'https://live.blockcypher.com/btc-testnet/',
   explorerAddressLink: 'https://live.blockcypher.com/btc-testnet/address/',
@@ -40,12 +41,12 @@ export const tBitcoin: Asset = {
 }
 
 export const atom: Asset = {
-  assetId: 'cosmos:cosmoshub-4/slip44:118',
-  chainId: 'cosmos:cosmoshub-4',
+  assetId: caip.cosmosAssetId,
+  chainId: caip.cosmosChainId,
   symbol: 'ATOM',
   name: 'Cosmos',
   precision: 6,
-  color: '#FFFFFF',
+  color: '#303F9F',
   icon: 'https://assets.coincap.io/assets/icons/256/atom.png',
   explorer: 'https://www.mintscan.io/cosmos',
   explorerAddressLink: 'https://www.mintscan.io/cosmos/account/',
@@ -53,14 +54,27 @@ export const atom: Asset = {
 }
 
 export const osmosis: Asset = {
-  assetId: 'cosmos:osmosis-1/slip44:118',
-  chainId: 'cosmos:osmosis-1',
+  assetId: caip.osmosisAssetId,
+  chainId: caip.osmosisChainId,
   symbol: 'OSMO',
   name: 'Osmosis',
   precision: 6,
-  color: '#FFFFFF',
+  color: '#750BBB',
   icon: 'https://assets.coincap.io/assets/icons/osmo@2x.png',
   explorer: 'https://www.mintscan.io/osmosis',
   explorerAddressLink: 'https://www.mintscan.io/osmosis/account/',
   explorerTxLink: 'https://www.mintscan.io/osmosis/txs/'
+}
+
+export const avax: Asset = {
+  assetId: caip.avalancheAssetId,
+  chainId: caip.avalancheChainId,
+  name: 'Avalanche',
+  symbol: 'AVAX',
+  precision: 18,
+  color: '#FFFFFF', // this will get picked up by the color generation script,
+  icon: 'https://rawcdn.githack.com/trustwallet/assets/32e51d582a890b3dd3135fe3ee7c20c2fd699a6d/blockchains/avalanchec/info/logo.png',
+  explorer: 'https://snowtrace.io',
+  explorerAddressLink: 'https://snowtrace.io/address/',
+  explorerTxLink: 'https://snowtrace.io/tx/'
 }