+++ packages/asset-service/src/generateAssetData/ethereum/yearnVaults.ts
@@ -1,30 +1,38 @@
+import { JsonRpcProvider } from '@ethersproject/providers'
 import { ethChainId as chainId, toAssetId } from '@shapeshiftoss/caip'
 import { Asset } from '@shapeshiftoss/types'
 import { Token, Vault } from '@yfi/sdk'
+import { Yearn } from '@yfi/sdk'
 import toLower from 'lodash/toLower'
 
-import { yearnSdk } from './yearnSdk'
+import { ethereum } from '../baseAssets'
+import { colorMap } from '../colorMap'
+
+const network = 1 // 1 for mainnet
+const provider = new JsonRpcProvider(process.env.REACT_APP_ETHEREUM_NODE_URL)
+export const yearnSdk = new Yearn(network, { provider })
+
+const explorerData = {
+  explorer: ethereum.explorer,
+  explorerAddressLink: ethereum.explorerAddressLink,
+  explorerTxLink: ethereum.explorerTxLink
+}
 
 export const getYearnVaults = async (): Promise<Asset[]> => {
   const vaults: Vault[] = await yearnSdk.vaults.get()
 
   return vaults.map((vault: Vault) => {
+    const assetId = toAssetId({ chainId, assetNamespace: 'erc20', assetReference: vault.address })
     return {
-      color: '#276BDB', // yearn.finance blue
+      color: colorMap[assetId] ?? '#FFFFFF',
       icon: vault.metadata.displayIcon,
       name: vault.name,
       precision: Number(vault.decimals),
       symbol: vault.symbol,
       tokenId: toLower(vault.address),
       chainId,
-      assetId: toAssetId({
-        chainId,
-        assetNamespace: 'erc20',
-        assetReference: vault.address
-      }),
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/'
+      assetId,
+      ...explorerData
     }
   })
 }
@@ -32,21 +40,17 @@ export const getYearnVaults = async (): Promise<Asset[]> => {
 export const getIronBankTokens = async (): Promise<Asset[]> => {
   const ironBankTokens: Token[] = await yearnSdk.ironBank.tokens()
   return ironBankTokens.map((token: Token) => {
+    const assetId = toAssetId({ chainId, assetNamespace: 'erc20', assetReference: token.address })
+
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
-      color: '#276BDB', // yearn.finance blue
+      ...explorerData,
+      color: colorMap[assetId] ?? '#FFFFFF',
       icon: token.icon ?? '',
       name: token.name,
       precision: Number(token.decimals),
       symbol: token.symbol,
       chainId,
-      assetId: toAssetId({
-        chainId,
-        assetNamespace: 'erc20',
-        assetReference: token.address
-      })
+      assetId
     }
   })
 }
@@ -54,21 +58,17 @@ export const getIronBankTokens = async (): Promise<Asset[]> => {
 export const getZapperTokens = async (): Promise<Asset[]> => {
   const zapperTokens: Token[] = await yearnSdk.tokens.supported()
   return zapperTokens.map((token: Token) => {
+    const assetId = toAssetId({ chainId, assetNamespace: 'erc20', assetReference: token.address })
+
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
-      color: '#7057F5', // zapper protocol purple
+      ...explorerData,
+      color: colorMap[assetId] ?? '#FFFFFF',
       icon: token.icon ?? '',
       name: token.name,
       precision: Number(token.decimals),
       symbol: token.symbol,
       chainId,
-      assetId: toAssetId({
-        chainId,
-        assetNamespace: 'erc20',
-        assetReference: token.address
-      })
+      assetId
     }
   })
 }
@@ -76,21 +76,17 @@ export const getZapperTokens = async (): Promise<Asset[]> => {
 export const getUnderlyingVaultTokens = async (): Promise<Asset[]> => {
   const underlyingTokens: Token[] = await yearnSdk.vaults.tokens()
   return underlyingTokens.map((token: Token) => {
+    const assetId = toAssetId({ chainId, assetNamespace: 'erc20', assetReference: token.address })
+
     return {
-      explorer: 'https://etherscan.io',
-      explorerAddressLink: 'https://etherscan.io/address/',
-      explorerTxLink: 'https://etherscan.io/tx/',
-      color: '#FFFFFF',
+      ...explorerData,
+      color: colorMap[assetId] ?? '#FFFFFF',
       icon: token.icon ?? '',
       name: token.name,
       precision: Number(token.decimals),
       symbol: token.symbol,
       chainId,
-      assetId: toAssetId({
-        chainId,
-        assetNamespace: 'erc20',
-        assetReference: token.address
-      })
+      assetId
     }
   })
 }