+++ packages/swapper/src/manager/SwapperManager.test.ts
@@ -7,6 +7,7 @@ import { SwapperType } from '../api'
 import { ThorchainSwapper, ZrxSwapper, ZrxSwapperDeps } from '../swappers'
 import { CowSwapper, CowSwapperDeps } from '../swappers/cow/CowSwapper'
 import { ThorchainSwapperDeps } from '../swappers/thorchain/types'
+import { WETH } from '../swappers/utils/test-data/assets'
 import { SwapperManager } from './SwapperManager'
 
 describe('SwapperManager', () => {
@@ -18,7 +19,7 @@ describe('SwapperManager', () => {
     apiUrl: 'https://api.cow.fi/mainnet/api/',
     adapter: <ethereum.ChainAdapter>{},
     web3: <Web3>{},
-    assetService: <AssetService>{}
+    feeAsset: WETH
   }
 
   const thorchainSwapperDeps: ThorchainSwapperDeps = {